refactor(tests): Removed `tests` suffix and created `local/lib` dir
