+++ test/local/crypto/base32.js
