+++ test/local/crypto/scrypt.js
