+++ test/local/password_check.js
